+++ src/order.ts
@@ -1,4 +1,4 @@
-import { IParticipant } from './action';
+import { IParticipantAsPerson, IParticipantAsProject, IParticipantAsWebApplication } from './action';
 import { IPaymentService, ITotalPaymentDue } from './action/trade/pay';
 import {
     IAmount as IMoneyTransferAmount,
@@ -227,7 +227,8 @@ export interface IAcceptedOffer<T extends IItemOffered> extends IOfferOptimized4
  * 販売者
  */
 export interface ISeller {
-    project: { id: string; typeOf: OrganizationType.Project };
+    // 最適化(2022-11-15~)
+    // project: { id: string; typeOf: OrganizationType.Project };
     id: string;
     typeOf: OrganizationType.Corporation;
     name: string;
@@ -235,21 +236,26 @@ export interface ISeller {
 /**
  * ウェブアプリケーションとしてのカスタマー
  */
-export type IWebApplicationCustomer = IWebApplication & IProfile;
+export type IWebApplicationCustomer = Pick<IWebApplication, 'id' | 'typeOf'> & IProfile;
 /**
  * 顧客組織としてのカスタマー
  */
-export type IOrganizationCustomer = ICustomerOrganization & IProfile;
+export type IOrganizationCustomer = Pick<ICustomerOrganization, 'id' | 'typeOf'> & IProfile;
+export type IPersonCustomer = Pick<IPerson, 'id' | 'typeOf' | 'memberOf'> & IProfile;
 /**
  * カスタマー
  */
-export type ICustomer = IPerson | IWebApplicationCustomer | IOrganizationCustomer;
-export type IBroker = IPerson;
+export type ICustomer = IPersonCustomer | IWebApplicationCustomer | IOrganizationCustomer;
+// 最適化(2022-11-15~)
+export type IBroker = Pick<IPerson, 'id' | 'identifier' | 'typeOf'>;
 /**
  * 返品者
  */
-export type IReturner = IParticipant;
+export type IParticipantAsReturner = IParticipantAsPerson | IParticipantAsProject | IParticipantAsWebApplication;
+// 最適化(2022-11-15~)
+export type IReturner = Pick<IParticipantAsReturner, 'id' | 'typeOf'>;
 export type IIdentifier = IPropertyValue<string>[];
+export type ISimpleCustomer = Pick<ICustomer, 'id' | 'typeOf' | 'name'>;
 export interface ISimpleOrder extends Pick<IThing, 'name'> {
     /**
      * object type
@@ -262,7 +268,7 @@ export interface ISimpleOrder extends Pick<IThing, 'name'> {
     /**
      * Party placing the order.
      */
-    customer: ICustomer;
+    customer: ISimpleCustomer;
     /**
      * A number that confirms the given order or payment has been received.
      */
@@ -335,6 +341,7 @@ export interface IOrder extends ISimpleOrder {
      * In most cases a broker never acquires or releases ownership of a product or service involved in an exchange.
      */
     broker?: IBroker;
+    customer: ICustomer;
     /**
      * Date order was returned.
      */