+++ CHANGELOG.md
@@ -18,6 +18,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Security
 
+## v4.277.0 - 2022-11-14
+
+### Changed
+
+- ISimpleOrderを最適化
+- 座席のIPlaceWithOfferを最適化
+- 注文のISeller,IOrganizationCustomer,IReturner,IBrokerを最適化
+- 所有権のIAcquiredFromを最適化
+- 取引のIAgentを最適化
+
 ## v4.276.0 - 2022-11-10
 
 ### Changed